+++ dca/tl_news_archive.php
@@ -14,6 +14,12 @@
 
 
 /**
+ * Register the global callbacks
+ */
+$GLOBALS['TL_DCA']['tl_news_archive']['config']['onload_callback'][] = array('tl_news_archive_categories', 'checkPermission');
+
+
+/**
  * Add a global operation to tl_news_archive
  */
 array_insert($GLOBALS['TL_DCA']['tl_news_archive']['list']['global_operations'], 1, array
@@ -26,3 +32,21 @@ array_insert($GLOBALS['TL_DCA']['tl_news_archive']['list']['global_operations'],
 		'attributes'          => 'onclick="Backend.getScrollOffset()" accesskey="c"'
 	)
 ));
+
+
+class tl_news_archive_categories extends Backend
+{
+
+	/**
+	 * Check the permission
+	 */
+	public function checkPermission()
+	{
+		$this->import('BackendUser', 'User');
+
+		if (!$this->User->isAdmin && !$this->User->newscategories)
+		{
+			unset($GLOBALS['TL_DCA']['tl_news_archive']['list']['global_operations']['categories']);
+		}
+	}
+}