+++ dca/tl_user.php
@@ -1,0 +1,32 @@
+<?php
+
+/**
+ * news_categories extension for Contao Open Source CMS
+ *
+ * Copyright (C) 2013 Codefog Ltd
+ *
+ * @package news_categories
+ * @author  Webcontext <http://webcontext.com>
+ * @author  Codefog Ltd <info@codefog.pl>
+ * @author  Kamil Kuzminski <kamil.kuzminski@codefog.pl>
+ * @license LGPL
+ */
+
+
+/**
+ * Add a palette to tl_user
+ */
+$GLOBALS['TL_DCA']['tl_user']['palettes']['extend'] = str_replace('newsfeedp;', 'newsfeedp,newscategories;', $GLOBALS['TL_DCA']['tl_user']['palettes']['extend']);
+$GLOBALS['TL_DCA']['tl_user']['palettes']['custom'] = str_replace('newsfeedp;', 'newsfeedp,newscategories;', $GLOBALS['TL_DCA']['tl_user']['palettes']['custom']);
+
+
+/**
+ * Add a new field to tl_user
+ */
+$GLOBALS['TL_DCA']['tl_user']['fields']['newscategories'] = array
+(
+	'label'                   => &$GLOBALS['TL_LANG']['tl_user']['newscategories'],
+	'exclude'                 => true,
+	'inputType'               => 'checkbox',
+	'sql'                     => "char(1) NOT NULL default ''"
+);