+++ languages/pl/tl_user.php
@@ -1,0 +1,19 @@
+<?php
+
+/**
+ * news_categories extension for Contao Open Source CMS
+ *
+ * Copyright (C) 2013 Codefog Ltd
+ *
+ * @package news_categories
+ * @author  Webcontext <http://webcontext.com>
+ * @author  Codefog Ltd <info@codefog.pl>
+ * @author  Kamil Kuzminski <kamil.kuzminski@codefog.pl>
+ * @license LGPL
+ */
+
+
+/**
+ * Fields
+ */
+$GLOBALS['TL_LANG']['tl_user']['newscategories'] = array('Zarządzanie kategoriami aktualności', 'Pozwól użytkownikowi na zarządzanie kategoriami aktualności.');